Time: 0 ms (100.00%), Space: 5.9 MB (23.05%) - LeetHub

+++ 374-guess-number-higher-or-lower/374-guess-number-higher-or-lower.cpp
@@ -12,7 +12,7 @@ public:
     int guessNumber(int n) {
         int start = 0;
         int end = n;
-       unsigned long long int mid;
+       int mid;
         while(start<=end){
             mid = start+(end - start)/2; 
             int pick = guess(mid);
